refactor(connect): move ripple method definitions to corresponding folder
