+++ packages/connect-common/src/types/api/ripple/index.ts
@@ -1,35 +1,12 @@
-import { MessagesSchema as PROTO } from '@trezor/protobuf';
 import type { Static } from '@trezor/schema-utils';
 import { Type } from '@trezor/schema-utils';
 
-import { DerivationPath } from '../../params';
+import type { rippleGetAddress } from './rippleGetAddress';
+import type { rippleSignTransaction } from './rippleSignTransaction';
 
-export type RipplePayment = Static<typeof RipplePayment>;
-export const RipplePayment = Type.Object({
-    amount: Type.String(),
-    destination: Type.String(),
-    destinationTag: Type.Optional(Type.Number()),
+// Ripple-specific operations
+export const TrezorConnectRipple = Type.Object({
+    rippleGetAddress: Type.Unsafe<typeof rippleGetAddress>(),
+    rippleSignTransaction: Type.Unsafe<typeof rippleSignTransaction>(),
 });
-
-export type RippleTransaction = Static<typeof RippleTransaction>;
-export const RippleTransaction = Type.Object({
-    fee: Type.String(),
-    flags: Type.Optional(Type.Number()),
-    sequence: Type.Number(),
-    maxLedgerVersion: Type.Optional(Type.Number()), // Proto: "last_ledger_sequence"
-    payment: RipplePayment,
-});
-
-export type RippleSignTransaction = Static<typeof RippleSignTransaction>;
-export const RippleSignTransaction = Type.Object({
-    path: DerivationPath,
-    transaction: RippleTransaction,
-    payment_req: Type.Optional(PROTO.PaymentRequest),
-    chunkify: Type.Optional(Type.Boolean()),
-});
-
-export type RippleSignedTx = Static<typeof RippleSignedTx>;
-export const RippleSignedTx = Type.Object({
-    serializedTx: Type.String(),
-    signature: Type.String(),
-});
+export type TrezorConnectRipple = Static<typeof TrezorConnectRipple>;
