+++ packages/connect-common/src/types/api/ripple/rippleSignTransaction.ts
@@ -1,0 +1,6 @@
+import type { RippleSignTransaction, RippleSignedTx } from './common';
+import type { Params, Response } from '../../params';
+
+export declare function rippleSignTransaction(
+    params: Params<RippleSignTransaction>,
+): Response<RippleSignedTx>;
