--- packages/connect-common/src/types/api/rippleSignTransaction.ts
@@ -1,6 +1,0 @@
-import type { Params, Response } from '../params';
-import type { RippleSignTransaction, RippleSignedTx } from './ripple';
-
-export declare function rippleSignTransaction(
-    params: Params<RippleSignTransaction>,
-): Response<RippleSignedTx>;
